requirements.txt
mediapipe==0.10.21
numpy==2.2.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pyserial==3.5
